pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Dhritiman-AgSpert <dhritiman.agspert@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.2"}
sqlalchemy = "^2.0.21"
psycopg2 = "^2.9.8"
python-jose = "^3.3.0"
alembic = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
